Fixed. Apparently, I write really, really bad code.
One thing people should probably be aware of: I overloaded the Note constructor. It now has a rescale argument that defaults to 1. I needed to do that so I could create a note with such a ridiculous value that it would never be the optimal choice. (Apparently, if it existed, a MIDI value of 1000000 would correspond to E8333.) This shouldn't change any existing behavior, though.