I took the improvisation code and concatenated my percussion code. I then rewrote the setup function so it runs my code then returns. It looks like my code is working.
Currently, the percussion and melody do not play together. I've got the percussion set up so you put in when you want it to play, but the melody just randomly picks how it wants to subdivide each measure. I feel like the easiest way to make this work would be to have my percussion code subdivide into the LCM of the possible outputs (currently 4, but 12 if you want triplets, etc; could get really large, really fast) and then play the melody note on the appropriate subdivision. I might need to make minor changes so it doesn't do anything too ridiculous, and then I'd have to add in the melody code, but I think that's the way to do it.