More Cleaning Up
To display images, they need to be in a folder called "data" in the project folder. So we need to create a folder called "data" in the Percussion folder with the RoboClub logo I'm uploading. This file should be moved to the Percussion folder and replace the one that's already there.

I redid the code so that draw() now gets called every subbeat instead of every beat. This allows us to update the display more frequently.

I also set up the code so that notes get stored in an ArrayList and get sent out at the start of the next subbeat. Hopefully this means the display is on track with the notes. Unfortunately, I have no way to test whether that's actually working. If that's not but everything else is, set the global variable "offset" to 2 and it'll display a header saying the display is behind. If other stuff is wrong, just pull an older version.