Hacked together some completely unreadable code that should hopefully do exactly what the old code did (other than accounting for computer-robot delays).
On the bright side, I should now be able to write code to tell the computer to generate new beats under a given set of constraints, which would be a step toward improvising/composing beat-by-beat instead of note-by-note. But I'd like to wait on that until someone can figure out how to test this on a system that can play sounds.