Added SimpleSynth 64-bit port
SimpleSynth hasn't been updated since 2009, and we probably want to keep using it. Recompiled the source code for macOS 10.14 and fixed MOST deprecation issues. Still not a 100% future proof, but should be good for now. Not completely tested but seems to do basic functions.

P.S. I checked and we are permitted to distribute copies of the program (read: about simplesynth, in app).