Started work on a (buggy) Markov chain for chords
Sorry if I'm stepping on someone's toes here, but I was travelling and bored.

I've started implmenting the chord-based Markov chain in chordHMM. It's not a full HMM yet, as it doesn't do anything with the melody line. It also seems to have issues identifying chords. I'll see if I get time to look into why that's the case; my first thought is octave wrapping interacting pooly with chord detection, but I haven't had time to finish debugging this.