Needs more testing
Normalized beatProbsArr correctly I hope. TODO test with oboe or something. Seems to mostly stick to whatever transposition it starts with, but it's hard to tell if it's correct without known notes. (Also had to jury-rig the Processing pitch detector - the other one wasn't working for some reason on my computer...)